# api/openapi.yaml
openapi: 3.0.3
info:
  title: China Medical and Tourism API
  version: "1.0.0"
  description: >
    REST API for the China Medical and Tourism platform. Implements the conventions
    fixed in docs/03-architecture/03-backend-api-architecture.md and
    docs/PROJECT_CONTEXT.md §8: REST/JSON under /api/v1, bearer JWT auth,
    cursor-based pagination for high-growth lists, a consistent error shape, and
    Idempotency-Key support on endpoints with external side effects.

    Request/response schemas mirror database/prisma/schema.prisma (Phase 6). Every
    endpoint below traces to one or more functional requirements from
    docs/01-product-planning/05-functional-requirements.md (FR-*), referenced in each
    operation's description.

    This spec covers the endpoints needed to power the 59 screens already built in
    apps/web (see docs/02-ui-ux-design/) — it is scoped to what those screens need,
    not a speculative superset.
servers:
  - url: https://api.cmt-platform.com/v1
    description: Production (illustrative — no real deployment exists yet)
  - url: http://localhost:3001/v1
    description: Local development
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    CursorParam:
      name: cursor
      in: query
      required: false
      description: Opaque pagination cursor from a previous response's nextCursor.
      schema:
        type: string
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ApplicationIdParam:
      name: applicationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    Forbidden:
      description: >
        Authenticated but not authorized for this resource (wrong role, or wrong
        tenant scope per BR-03/BR-04).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: { code: FORBIDDEN, message: "You do not have access to this resource." }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: { code: NOT_FOUND, message: "The requested resource was not found." }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable identifier, e.g. CASE_NOT_FOUND.
            message:
              type: string
            details:
              type: object
    PaginatedMeta:
      type: object
      properties:
        nextCursor:
          type: string
          nullable: true
        hasMore:
          type: boolean
...
    Hospital:
      type: object
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        citySlug: { type: string }
        description: { type: string }
        priceTier: { type: string, enum: ["$$", "$$$", "$$$$"] }
        accreditations: { type: array, items: { type: string } }
        languages: { type: array, items: { type: string } }
        facilities: { type: array, items: { type: string } }
        rating: { type: number, format: float }
        reviewCount: { type: integer }
        status: { type: string, enum: [Draft, PendingApproval, Published, Rejected] }
    Doctor:
      type: object
      properties:
        id: { type: string, format: uuid }
        hospitalId: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        specialtySlug: { type: string }
        credentials: { type: string }
        yearsExperience: { type: integer }
        languages: { type: array, items: { type: string } }
        bio: { type: string }
    TreatmentPackage:
      type: object
      properties:
        id: { type: string, format: uuid }
        hospitalId: { type: string, format: uuid }
        name: { type: string }
        specialtySlug: { type: string }
        description: { type: string }
        priceMinUsd: { type: number }
        priceMaxUsd: { type: number }
        includes: { type: array, items: { type: string } }
        status: { type: string, enum: [Draft, Active, Archived] }
...
    TransferRequest:
      type: object
      properties:
        id: { type: string, format: uuid }
        applicationId: { type: string, format: uuid }
        direction: { type: string, enum: [Arrival, Departure] }
        flightNumber: { type: string, nullable: true }
        scheduledAt: { type: string, format: date-time }
        pickupLocation: { type: string }
        driverId: { type: string, format: uuid, nullable: true }
        status: { type: string, enum: [Requested, Assigned, Completed, Cancelled] }
    CreateTransferRequest:
      type: object
      required: [direction, scheduledAt, pickupLocation]
      properties:
        direction: { type: string, enum: [Arrival, Departure] }
        flightNumber: { type: string }
        scheduledAt: { type: string, format: date-time }
        pickupLocation: { type: string }
    InterpreterAssignment:
      type: object
      properties:
        id: { type: string, format: uuid }
        applicationId: { type: string, format: uuid }
        interpreterId: { type: string, format: uuid, nullable: true }
        hospitalVisitAt: { type: string, format: date-time }
        department: { type: string, nullable: true }
        note: { type: string, nullable: true }
        status: { type: string, enum: [Requested, Assigned, Completed, Cancelled] }
...
    PlatformAnalytics:
      type: object
      properties:
        bookingsThisMonth: { type: integer }
        revenueThisMonthUsd: { type: number }
        activeCases: { type: integer }
        slaComplianceRate: { type: number }
        conversionFunnel:
          type: array
          items:
            type: object
            properties:
              stage: { type: string }
              count: { type: integer }
paths:
...
  /hospitals:
    get:
      operationId: getHospitals
      tags: [Hospitals]
      summary: Search/browse published hospitals (FR-HOSP-06)
      security: []
      parameters:
        - name: specialty
          in: query
          schema: { type: string }
        - name: city
          in: query
          schema: { type: string }
        - name: language
          in: query
          schema: { type: string }
        - name: minRating
          in: query
          schema: { type: number }
        - { $ref: "#/components/parameters/CursorParam" }
        - { $ref: "#/components/parameters/LimitParam" }
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Hospital" }
                  meta: { $ref: "#/components/schemas/PaginatedMeta" }
...
  /hospitals/{hospitalId}:
    parameters:
      - name: hospitalId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getHospitalsByHospitalId
      tags: [Hospitals]
      summary: Get hospital detail (public if status=Published)
      security: []
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Hospital" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: patchHospitalsByHospitalId
      tags: [Hospitals]
      summary: Update hospital profile (hospital_staff only, scoped to own hospital per BR-03; submits a HospitalModerationItem rather than a live edit, per FR-HOSP-05)
      responses:
        "202": { description: Change submitted for Admin approval. }
        "403": { $ref: "#/components/responses/Forbidden" }
  /hospitals/{hospitalId}/doctors:
    parameters:
      - name: hospitalId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getHospitalsByHospitalIdDoctors
      tags: [Hospitals]
      summary: List a hospital's doctors
      security: []
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Doctor" }
    post:
      operationId: postHospitalsByHospitalIdDoctors
      tags: [Hospitals]
      summary: Add a doctor (hospital_staff, own hospital) (FR-HOSP-03)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Doctor" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Doctor" }
  /hospitals/{hospitalId}/doctors/{doctorId}:
    parameters:
      - name: hospitalId
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: doctorId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getHospitalsByHospitalIdDoctorsByDoctorId
      tags: [Hospitals]
      summary: Get doctor detail
      security: []
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Doctor" }
  /hospitals/{hospitalId}/packages:
    parameters:
      - name: hospitalId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getHospitalsByHospitalIdPackages
      tags: [Hospitals]
      summary: List a hospital's treatment packages (FR-HOSP-04)
      security: []
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TreatmentPackage" }
    post:
      operationId: postHospitalsByHospitalIdPackages
      tags: [Hospitals]
      summary: Create a treatment package (hospital_staff, own hospital)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TreatmentPackage" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TreatmentPackage" }
  /hospitals/{hospitalId}/reports:
    parameters:
      - name: hospitalId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getHospitalsByHospitalIdReports
      tags: [Hospitals]
      summary: Hospital-scoped booking/revenue/conversion report (FR-ANALYTICS-02, Screen 43)
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlatformAnalytics" }
        "403": { $ref: "#/components/responses/Forbidden" }

  # -------------------------------------------------------------------------
  # Applications — FR-APP-01 through FR-APP-08
  # -------------------------------------------------------------------------
...
  /applications/{applicationId}/transfers:
    parameters:
      - { $ref: "#/components/parameters/ApplicationIdParam" }
    get:
      operationId: getApplicationsByApplicationIdTransfers
      tags: [Transport]
      summary: List transfer requests for a case
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TransferRequest" }
    post:
      operationId: postApplicationsByApplicationIdTransfers
      tags: [Transport]
      summary: Request an airport transfer (patient) (FR-XFER-01)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTransferRequest" }
      responses:
        "201":
          description: Requested.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TransferRequest" }
  /transfers/{transferId}/assign:
    parameters:
      - name: transferId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: postTransfersByTransferIdAssign
      tags: [Transport]
      summary: Assign a driver (case_manager/admin only) (FR-XFER-02, BR-23)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [driverId]
              properties:
                driverId: { type: string, format: uuid }
      responses:
        "200": { description: Assigned; driver and patient notified. }
  /transfers/{transferId}/complete:
    parameters:
      - name: transferId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: postTransfersByTransferIdComplete
      tags: [Transport]
      summary: Mark a transfer complete (driver only, own assignment) (FR-XFER-04)
      responses:
        "200": { description: Completed; unlocks driver payment eligibility per BR-24. }
...
  /drivers/me/trips:
    get:
      operationId: getDriversMeTrips
      tags: [Transport]
      summary: List the authenticated driver's assigned trips (FR-XFER-03, Screen 58)
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [Assigned, Completed] }
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TransferRequest" }
  /applications/{applicationId}/interpreter-sessions:
    parameters:
      - { $ref: "#/components/parameters/ApplicationIdParam" }
    get:
      operationId: getApplicationsByApplicationIdInterpreterSessions
      tags: [Transport]
      summary: List interpreter sessions for a case
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/InterpreterAssignment" }
    post:
      operationId: postApplicationsByApplicationIdInterpreterSessions
      tags: [Transport]
      summary: Request an interpreter session (case_manager only) (FR-INTERP-01)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hospitalVisitAt]
              properties:
                hospitalVisitAt: { type: string, format: date-time }
                department: { type: string }
                note: { type: string }
      responses:
        "201": { description: Requested. }
  /interpreter-sessions/{sessionId}/assign:
    parameters:
      - name: sessionId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: postInterpreterSessionsBySessionIdAssign
      tags: [Transport]
      summary: Assign an interpreter (case_manager/admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [interpreterId]
              properties:
                interpreterId: { type: string, format: uuid }
      responses:
        "200": { description: Assigned. }
  /interpreter-sessions/{sessionId}/complete:
    parameters:
      - name: sessionId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: postInterpreterSessionsBySessionIdComplete
      tags: [Transport]
      summary: Mark an interpreter session complete (interpreter only, own assignment)
      responses:
        "200": { description: Completed. }
...